refactor: mv username module
