+++ index.js
@@ -1,6 +1,6 @@
 import { createInterface } from 'node:readline/promises';
 import { stdin as input, stdout as output } from 'node:process';
-import { username } from './src/username.js';
+import { username } from './src/utils/username.js';
 import { printCWD } from './src/utils/navigator.js';
 import { processInput } from './src/input-processing.js';
 import { blue, yellow } from './src/utils/logger.js';
